+++ src/server/game/Server/Packets/NPCPackets.h
@@ -80,7 +80,7 @@ namespace WorldPackets
             int32 GossipOptionID  = 0;
             GossipOptionNpc OptionNPC = {};
             uint8 OptionFlags   = 0;
-            int32 OptionCost    = 0;
+            uint64 OptionCost    = 0;
             uint32 OptionLanguage = 0;
             GossipOptionFlags Flags = {};
             int32 OrderIndex = 0;
